Move configs to own addon
