+++ addons/bombs/CfgMagazines.hpp
@@ -1,0 +1,3 @@
+class CfgMagazines {
+
+};
