+++ addons/bombs/CfgWeapons.hpp
@@ -1,0 +1,9 @@
+class CfgWeapons {
+    class RocketPods;
+    class Bomb_04_Plane_CAS_01_F: RocketPods {
+        ace_laser_canSelect = 1;
+        ace_laser_showHud = 1;
+        canLock = 0;
+        weaponLockSystem = 0;
+    };
+};
