+++ addons/missileguidance/CfgWeapons.hpp
@@ -6,12 +6,4 @@ class CfgWeapons {
         lockingTargetSound[] = {"",0,1};
         lockedTargetSound[] = {"",0,1};
     };
-    
-    class RocketPods;
-    class Bomb_04_Plane_CAS_01_F: RocketPods {
-        ace_laser_canSelect = 1;
-        ace_laser_showHud = 1;
-        canLock = 0;
-        weaponLockSystem = 0;
-    };
 };
